Rewrite database.py and renamed in database_hw3.py.
